v.6.2.1 perbaikan bug pada fitur login multi user

+++ app/Http/Controllers/GoodsCRUDcontrol.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Food;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class GoodsCRUDcontrol extends Controller
 {
@@ -14,36 +15,44 @@ class GoodsCRUDcontrol extends Controller
      */
     // admin
         public function aindex(){
-            $makanan = Food::all();
-            $hitung_makanan_tersedia = Food::where('jenis_masakan', 'food')->where('status_masakan', 'available')->count();
-            $hitung_makanan_habis = Food::where('jenis_masakan', 'food')->where('status_masakan', 'run out')->count();
-            $hitung_minuman_tersedia = Food::where('jenis_masakan', 'drink')->where('status_masakan', 'available')->count();
-            $hitung_minuman_habis = Food::where('jenis_masakan', 'drink')->where('status_masakan', 'run out')->count();
-            $data = [
-                'makanan' => $makanan,
-                'hitung_makanan_tersedia' => $hitung_makanan_tersedia,
-                'hitung_makanan_habis' => $hitung_makanan_habis,
-                'hitung_minuman_tersedia' => $hitung_minuman_tersedia,
-                'hitung_minuman_habis' => $hitung_minuman_habis
-            ];
-            return view('admin/m_goods', $data);
+            if (Auth::user()->id_level == 1) {
+                $makanan = Food::all();
+                $hitung_makanan_tersedia = Food::where('jenis_masakan', 'food')->where('status_masakan', 'available')->count();
+                $hitung_makanan_habis = Food::where('jenis_masakan', 'food')->where('status_masakan', 'run out')->count();
+                $hitung_minuman_tersedia = Food::where('jenis_masakan', 'drink')->where('status_masakan', 'available')->count();
+                $hitung_minuman_habis = Food::where('jenis_masakan', 'drink')->where('status_masakan', 'run out')->count();
+                $data = [
+                    'makanan' => $makanan,
+                    'hitung_makanan_tersedia' => $hitung_makanan_tersedia,
+                    'hitung_makanan_habis' => $hitung_makanan_habis,
+                    'hitung_minuman_tersedia' => $hitung_minuman_tersedia,
+                    'hitung_minuman_habis' => $hitung_minuman_habis
+                ];
+                return view('admin/m_goods', $data);
+            }else{
+                return redirect()->back();
+            }
         }
 
-    // waiter
+        // waiter
         public function windex(){
-            $makanan = Food::all();
-            $hitung_makanan_tersedia = Food::where('jenis_masakan', 'food')->where('status_masakan', 'available')->count();
-            $hitung_makanan_habis = Food::where('jenis_masakan', 'food')->where('status_masakan', 'run out')->count();
-            $hitung_minuman_tersedia = Food::where('jenis_masakan', 'drink')->where('status_masakan', 'available')->count();
-            $hitung_minuman_habis = Food::where('jenis_masakan', 'drink')->where('status_masakan', 'run out')->count();
-            $data = [
-                'makanan' => $makanan,
-                'hitung_makanan_tersedia' => $hitung_makanan_tersedia,
-                'hitung_makanan_habis' => $hitung_makanan_habis,
-                'hitung_minuman_tersedia' => $hitung_minuman_tersedia,
-                'hitung_minuman_habis' => $hitung_minuman_habis
-            ];
-            return view('waiter/m_goods', $data);
+            if (Auth::user()->id_level == 2) {
+                $makanan = Food::all();
+                $hitung_makanan_tersedia = Food::where('jenis_masakan', 'food')->where('status_masakan', 'available')->count();
+                $hitung_makanan_habis = Food::where('jenis_masakan', 'food')->where('status_masakan', 'run out')->count();
+                $hitung_minuman_tersedia = Food::where('jenis_masakan', 'drink')->where('status_masakan', 'available')->count();
+                $hitung_minuman_habis = Food::where('jenis_masakan', 'drink')->where('status_masakan', 'run out')->count();
+                $data = [
+                    'makanan' => $makanan,
+                    'hitung_makanan_tersedia' => $hitung_makanan_tersedia,
+                    'hitung_makanan_habis' => $hitung_makanan_habis,
+                    'hitung_minuman_tersedia' => $hitung_minuman_tersedia,
+                    'hitung_minuman_habis' => $hitung_minuman_habis
+                ];
+                return view('waiter/m_goods', $data);
+            }else{
+                return redirect()->back();
+            }
         }
 
     /**
@@ -53,12 +62,20 @@ class GoodsCRUDcontrol extends Controller
      */
     // admin
         public function acreate(){
-            return view('admin/t_goods');
+            if (Auth::user()->id_level == 1) {
+                return view('admin/t_goods');
+            }else{
+                return redirect()->back();
+            }
         }
 
     // waiter
         public function wcreate(){
-            return view('waiter/t_goods');
+            if (Auth::user()->id_level == 2) {
+                return view('waiter/t_goods');
+            }else{
+                return redirect()->back();
+            }
         }
 
     /**
@@ -69,58 +86,66 @@ class GoodsCRUDcontrol extends Controller
      */
     // admin
         public function astore(Request $request){
-            // validation
-                $rule_message = [
-                    'namamasakan.required'=>'You cant leave Food Name field empty',
-                    'tipemasakan.required'=>'You cant leave Type of Food field empty',
-                    'hargamasakan.required'=>'You cant leave Food Price field empty',
-                    'statusmasakan.required'=>'You cant leave Food Status field empty'
-                ];
-                $rules = [
-                        'namamasakan' => 'required',
-                        'tipemasakan' => 'required',
-                        'hargamasakan' => 'required',
-                        'statusmasakan' => 'required'
+            if (Auth::user()->id_level == 1) {
+                // validation
+                    $rule_message = [
+                        'namamasakan.required'=>'You cant leave Food Name field empty',
+                        'tipemasakan.required'=>'You cant leave Type of Food field empty',
+                        'hargamasakan.required'=>'You cant leave Food Price field empty',
+                        'statusmasakan.required'=>'You cant leave Food Status field empty'
                     ];
+                    $rules = [
+                            'namamasakan' => 'required',
+                            'tipemasakan' => 'required',
+                            'hargamasakan' => 'required',
+                            'statusmasakan' => 'required'
+                        ];
 
-                $this->validate($request, $rules, $rule_message);
+                    $this->validate($request, $rules, $rule_message);
 
-            Food::create([
-                'nama_masakan' => $request->namamasakan,
-                'jenis_masakan' => $request->tipemasakan,
-                'harga' => $request->hargamasakan,
-                'status_masakan' => $request->statusmasakan
-            ]);
+                Food::create([
+                    'nama_masakan' => $request->namamasakan,
+                    'jenis_masakan' => $request->tipemasakan,
+                    'harga' => $request->hargamasakan,
+                    'status_masakan' => $request->statusmasakan
+                ]);
 
-            return redirect('/adashboard/goods')->with('success', "Data ->{$request->namamasakan}<- was successfully added !");
+                return redirect('/adashboard/goods')->with('success', "Data ->{$request->namamasakan}<- was successfully added !");
+            }else{
+                return redirect()->back();
+            }
         }
 
     // waiter
         public function wstore(Request $request){
-            // validation
-                $rule_message = [
-                    'namamasakan.required'=>'You cant leave Food Name field empty',
-                    'tipemasakan.required'=>'You cant leave Type of Food field empty',
-                    'hargamasakan.required'=>'You cant leave Food Price field empty',
-                    'statusmasakan.required'=>'You cant leave Food Status field empty'
-                ];
-                $rules = [
-                        'namamasakan' => 'required',
-                        'tipemasakan' => 'required',
-                        'hargamasakan' => 'required',
-                        'statusmasakan' => 'required'
+            if (Auth::user()->id_level == 2) {
+                // validation
+                    $rule_message = [
+                        'namamasakan.required'=>'You cant leave Food Name field empty',
+                        'tipemasakan.required'=>'You cant leave Type of Food field empty',
+                        'hargamasakan.required'=>'You cant leave Food Price field empty',
+                        'statusmasakan.required'=>'You cant leave Food Status field empty'
                     ];
+                    $rules = [
+                            'namamasakan' => 'required',
+                            'tipemasakan' => 'required',
+                            'hargamasakan' => 'required',
+                            'statusmasakan' => 'required'
+                        ];
 
-                $this->validate($request, $rules, $rule_message);
+                    $this->validate($request, $rules, $rule_message);
 
-            Food::create([
-                'nama_masakan' => $request->namamasakan,
-                'jenis_masakan' => $request->tipemasakan,
-                'harga' => $request->hargamasakan,
-                'status_masakan' => $request->statusmasakan
-            ]);
+                Food::create([
+                    'nama_masakan' => $request->namamasakan,
+                    'jenis_masakan' => $request->tipemasakan,
+                    'harga' => $request->hargamasakan,
+                    'status_masakan' => $request->statusmasakan
+                ]);
 
-            return redirect('/wdashboard/goods')->with('success', "Data ->{$request->namamasakan}<- was successfully added !");
+                return redirect('/wdashboard/goods')->with('success', "Data ->{$request->namamasakan}<- was successfully added !");
+            }else{
+                return redirect()->back();
+            }
         }
 
     /**
@@ -131,16 +156,24 @@ class GoodsCRUDcontrol extends Controller
      */
     // admin
         public function ashow(Food $food){
-            $tipemasakan = ['food', 'drink'];
-            $statusmasakan = ['available', 'run out'];
-            return view('admin/d_goods', ['makanan' => $food, 'tipemasakan' => $tipemasakan, 'statusmasakan' => $statusmasakan]);
+            if (Auth::user()->id_level == 1) {
+                $tipemasakan = ['food', 'drink'];
+                $statusmasakan = ['available', 'run out'];
+                return view('admin/d_goods', ['makanan' => $food, 'tipemasakan' => $tipemasakan, 'statusmasakan' => $statusmasakan]);
+            }else{
+                return redirect()->back();
+            }
         }
 
     // waiter
         public function wshow(Food $food){
-            $tipemasakan = ['food', 'drink'];
-            $statusmasakan = ['available', 'run out'];
-            return view('waiter/d_goods', ['makanan' => $food, 'tipemasakan' => $tipemasakan, 'statusmasakan' => $statusmasakan]);
+            if (Auth::user()->id_level == 2) {
+                $tipemasakan = ['food', 'drink'];
+                $statusmasakan = ['available', 'run out'];
+                return view('waiter/d_goods', ['makanan' => $food, 'tipemasakan' => $tipemasakan, 'statusmasakan' => $statusmasakan]);
+            }else{
+                return redirect()->back();
+            }
         }
     /**
      * Show the form for editing the specified resource.
@@ -151,17 +184,25 @@ class GoodsCRUDcontrol extends Controller
     // admin
         public function aedit(Food $food){
             // dd($food);
-            $tipemasakan = ['food', 'drink'];
-            $statusmasakan = ['available', 'run out'];
-            return view('admin/e_goods', ['makanan' => $food, 'tipemasakan' => $tipemasakan, 'statusmasakan' => $statusmasakan]);
+            if (Auth::user()->id_level == 1) {
+                $tipemasakan = ['food', 'drink'];
+                $statusmasakan = ['available', 'run out'];
+                return view('admin/e_goods', ['makanan' => $food, 'tipemasakan' => $tipemasakan, 'statusmasakan' => $statusmasakan]);
+            }else{
+                return redirect()->back();
+            }
         }
 
     // waiter
         public function wedit(Food $food){
             // dd($food);
-            $tipemasakan = ['food', 'drink'];
-            $statusmasakan = ['available', 'run out'];
-            return view('waiter/e_goods', ['makanan' => $food, 'tipemasakan' => $tipemasakan, 'statusmasakan' => $statusmasakan]);
+            if (Auth::user()->id_level == 2) {
+                $tipemasakan = ['food', 'drink'];
+                $statusmasakan = ['available', 'run out'];
+                return view('waiter/e_goods', ['makanan' => $food, 'tipemasakan' => $tipemasakan, 'statusmasakan' => $statusmasakan]);
+            }else{
+                return redirect()->back();
+            }
         }
 
     /**
@@ -173,36 +214,41 @@ class GoodsCRUDcontrol extends Controller
      */
     // admin
         public function aupdate(Request $request, Food $food){
-            // validation
-                $rule_message = [
-                    'namamasakan.required'=>'You cant leave Food Name field empty',
-                    'tipemasakan.required'=>'You cant leave Type of Food field empty',
-                    'hargamasakan.required'=>'You cant leave Food Price field empty',
-                    'statusmasakan.required'=>'You cant leave Food Status field empty'
-                ];
-                $rules = [
-                        'namamasakan' => 'required',
-                        'tipemasakan' => 'required',
-                        'hargamasakan' => 'required',
-                        'statusmasakan' => 'required'
+            if (Auth::user()->id_level == 1) {
+                // validation
+                    $rule_message = [
+                        'namamasakan.required'=>'You cant leave Food Name field empty',
+                        'tipemasakan.required'=>'You cant leave Type of Food field empty',
+                        'hargamasakan.required'=>'You cant leave Food Price field empty',
+                        'statusmasakan.required'=>'You cant leave Food Status field empty'
                     ];
+                    $rules = [
+                            'namamasakan' => 'required',
+                            'tipemasakan' => 'required',
+                            'hargamasakan' => 'required',
+                            'statusmasakan' => 'required'
+                        ];
 
-                $this->validate($request, $rules, $rule_message);
+                    $this->validate($request, $rules, $rule_message);
 
-            $data = [
-                'nama_masakan' => $request->namamasakan,
-                'jenis_masakan' => $request->tipemasakan,
-                'harga' => $request->hargamasakan,
-                'status_masakan' => $request->statusmasakan
-            ];
+                $data = [
+                    'nama_masakan' => $request->namamasakan,
+                    'jenis_masakan' => $request->tipemasakan,
+                    'harga' => $request->hargamasakan,
+                    'status_masakan' => $request->statusmasakan
+                ];
 
-            Food::where('id', $food->id)->update($data);
+                Food::where('id', $food->id)->update($data);
 
-            return redirect('/adashboard/goods')->with('success', "Data ->{$request->namamasakan}<- was successfully edited !");
+                return redirect('/adashboard/goods')->with('success', "Data ->{$request->namamasakan}<- was successfully edited !");
+            }else{
+                return redirect()->back();
+            }
         }
 
     // waiter
         public function wupdate(Request $request, Food $food){
+            if (Auth::user()->id_level == 2) {
             // validation
                 $rule_message = [
                     'namamasakan.required'=>'You cant leave Food Name field empty',
@@ -219,16 +265,19 @@ class GoodsCRUDcontrol extends Controller
 
                 $this->validate($request, $rules, $rule_message);
 
-            $data = [
-                'nama_masakan' => $request->namamasakan,
-                'jenis_masakan' => $request->tipemasakan,
-                'harga' => $request->hargamasakan,
-                'status_masakan' => $request->statusmasakan
-            ];
+                $data = [
+                    'nama_masakan' => $request->namamasakan,
+                    'jenis_masakan' => $request->tipemasakan,
+                    'harga' => $request->hargamasakan,
+                    'status_masakan' => $request->statusmasakan
+                ];
 
-            Food::where('id', $food->id)->update($data);
+                Food::where('id', $food->id)->update($data);
 
-            return redirect('/wdashboard/goods')->with('success', "Data ->{$request->namamasakan}<- was successfully edited !");
+                return redirect('/wdashboard/goods')->with('success', "Data ->{$request->namamasakan}<- was successfully edited !");
+            }else{
+                return redirect()->back();
+            }
         }
 
     /**
@@ -239,13 +288,21 @@ class GoodsCRUDcontrol extends Controller
      */
     // admin
         public function adestroy(Food $food){
-            $food->delete();
-            return redirect('/adashboard/goods')->with('success', "Data ->{$food->namamasakan}<- was successfully deleted !");
+            if (Auth::user()->id_level == 1) {
+                $food->delete();
+                return redirect('/adashboard/goods')->with('success', "Data ->{$food->namamasakan}<- was successfully deleted !");
+            }else{
+                return redirect()->back();
+            }
         }
 
     // waiter
         public function wdestroy(Food $food){
-            $food->delete();
-            return redirect('/wdashboard/goods')->with('success', "Data ->{$food->namamasakan}<- was successfully deleted !");
+            if (Auth::user()->id_level == 2) {
+                $food->delete();
+                return redirect('/wdashboard/goods')->with('success', "Data ->{$food->namamasakan}<- was successfully deleted !");
+            }else{
+                return redirect()->back();
+            }
         }
 }
