+++ app/Http/Controllers/UsersCRUDcontrol.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\User;
 use App\Level;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class UsersCRUDcontrol extends Controller
 {
@@ -13,25 +14,28 @@ class UsersCRUDcontrol extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
-    {
-        $user = User::all();
-        $level = Level::all();
-        $admin = User::where('id_level', 1)->count();
-        $waiter = User::where('id_level', 2)->count();
-        $kasir = User::where('id_level', 3)->count();
-        $owner = User::where('id_level', 4)->count();
+    public function index(){
+        if (Auth::user()->id_level == 1) {
+            $user = User::all();
+            $level = Level::all();
+            $admin = User::where('id_level', 1)->count();
+            $waiter = User::where('id_level', 2)->count();
+            $kasir = User::where('id_level', 3)->count();
+            $owner = User::where('id_level', 4)->count();
 
-        $data = [
-            'user' => $user,
-            'level' => $level,
-            'hitung_admin' => $admin,
-            'hitung_waiter' => $waiter,
-            'hitung_kasir' => $kasir,
-            'hitung_owner' => $owner
-        ];
+            $data = [
+                'user' => $user,
+                'level' => $level,
+                'hitung_admin' => $admin,
+                'hitung_waiter' => $waiter,
+                'hitung_kasir' => $kasir,
+                'hitung_owner' => $owner
+            ];
 
-        return view('admin/m_user', $data);
+            return view('admin/m_user', $data);
+        }else{
+            return redirect()->back();
+        }
     }
 
     /**
@@ -39,10 +43,13 @@ class UsersCRUDcontrol extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function create()
-    {
-        $level = Level::all();
-        return view('admin/t_user', ['level' => $level]);
+    public function create(){
+        if (Auth::user()->id_level == 1) {
+            $level = Level::all();
+            return view('admin/t_user', ['level' => $level]);
+        }else{
+            return redirect()->back();
+        }
     }
 
     /**
@@ -51,36 +58,39 @@ class UsersCRUDcontrol extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
-    {
-        // validation
-            $rule_message = [
-                'username.required'=>'You cant leave Username field empty',
-                'username.unique'=>'This username has been registered',
-                'password.required'=>'You cant leave Password field empty',
-                'name.required'=>'You cant leave Fullname field empty',
-                'genderRadios.required'=>'You cant leave Gender field empty',
-                'role.required' =>'You cant leave User Role field empty'
-            ];
-            $rules = [
-                    'username' => 'required|unique:users,username',
-                    'password' => 'required',
-                    'name' => 'required',
-                    'genderRadios' => 'required',
-                    'role' => 'required',
+    public function store(Request $request){
+        if (Auth::user()->id_level == 1) {
+            // validation
+                $rule_message = [
+                    'username.required'=>'You cant leave Username field empty',
+                    'username.unique'=>'This username has been registered',
+                    'password.required'=>'You cant leave Password field empty',
+                    'name.required'=>'You cant leave Fullname field empty',
+                    'genderRadios.required'=>'You cant leave Gender field empty',
+                    'role.required' =>'You cant leave User Role field empty'
                 ];
+                $rules = [
+                        'username' => 'required|unique:users,username',
+                        'password' => 'required',
+                        'name' => 'required',
+                        'genderRadios' => 'required',
+                        'role' => 'required',
+                    ];
 
-            $this->validate($request, $rules, $rule_message);
+                $this->validate($request, $rules, $rule_message);
 
-        User::create([
-            'username' => $request->username,
-            'password' => bcrypt($request->password),
-            'nama_user' => $request->name,
-            'jenis_kelamin' => $request->genderRadios,
-            'id_level' => $request->role
-        ]);
+            User::create([
+                'username' => $request->username,
+                'password' => bcrypt($request->password),
+                'nama_user' => $request->name,
+                'jenis_kelamin' => $request->genderRadios,
+                'id_level' => $request->role
+            ]);
 
-        return redirect('/adashboard/users')->with('success', "Data ->{$request->username}<- was successfully added !");
+            return redirect('/adashboard/users')->with('success', "Data ->{$request->username}<- was successfully added !");
+        }else{
+            return redirect()->back();
+        }
     }
 
     /**
@@ -89,10 +99,13 @@ class UsersCRUDcontrol extends Controller
      * @param  \App\User  $user
      * @return \Illuminate\Http\Response
      */
-    public function show(User $user)
-    {
-        $level = Level::all();
-        return view('admin/d_user', ['user' => $user, 'level' => $level]);
+    public function show(User $user){
+        if (Auth::user()->id_level == 1) {
+            $level = Level::all();
+            return view('admin/d_user', ['user' => $user, 'level' => $level]);
+        }else{
+            return redirect()->back();
+        }
     }
 
     /**
@@ -101,10 +114,13 @@ class UsersCRUDcontrol extends Controller
      * @param  \App\User  $user
      * @return \Illuminate\Http\Response
      */
-    public function edit(User $user)
-    {
-        $level = Level::all();
-        return view('admin/e_user', ['user' => $user, 'level' => $level]);
+    public function edit(User $user){
+        if (Auth::user()->id_level == 1) {
+            $level = Level::all();
+            return view('admin/e_user', ['user' => $user, 'level' => $level]);
+        }else{
+            return redirect()->back();
+        }
     }
 
     /**
@@ -114,38 +130,41 @@ class UsersCRUDcontrol extends Controller
      * @param  \App\User  $user
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, User $user)
-    {
-        // validation
-            $rule_message = [
-                'username.required'=>'You cant leave Username field empty',
-                'username.unique'=>'This username has been registered',
-                'password.required'=>'You cant leave Password field empty',
-                'name.required'=>'You cant leave Fullname field empty',
-                'genderRadios.required'=>'You cant leave Gender field empty',
-                'role.required' =>'You cant leave User Role field empty'
-            ];
-            $rules = [
-                    'username' => 'required|unique:users,username',
-                    'password' => 'required',
-                    'name' => 'required',
-                    'genderRadios' => 'required',
-                    'role' => 'required',
+    public function update(Request $request, User $user){
+        if (Auth::user()->id_level == 1) {
+            // validation
+                $rule_message = [
+                    'username.required'=>'You cant leave Username field empty',
+                    'username.unique'=>'This username has been registered',
+                    'password.required'=>'You cant leave Password field empty',
+                    'name.required'=>'You cant leave Fullname field empty',
+                    'genderRadios.required'=>'You cant leave Gender field empty',
+                    'role.required' =>'You cant leave User Role field empty'
                 ];
+                $rules = [
+                        'username' => 'required|unique:users,username',
+                        'password' => 'required',
+                        'name' => 'required',
+                        'genderRadios' => 'required',
+                        'role' => 'required',
+                    ];
 
-            $this->validate($request, $rules, $rule_message);
+                $this->validate($request, $rules, $rule_message);
 
-        $data = [
-            'username' => $request->username,
-            'password' => bcrypt($request->password),
-            'nama_user' => $request->name,
-            'jenis_kelamin' => $request->genderRadios,
-            'id_level' => $request->role
-        ];
+            $data = [
+                'username' => $request->username,
+                'password' => bcrypt($request->password),
+                'nama_user' => $request->name,
+                'jenis_kelamin' => $request->genderRadios,
+                'id_level' => $request->role
+            ];
 
-        User::where('id', $user->id)->update($data);
+            User::where('id', $user->id)->update($data);
 
-        return redirect('/adashboard/users')->with('success', "Data ->{$request->username}<- was successfully edited !");
+            return redirect('/adashboard/users')->with('success', "Data ->{$request->username}<- was successfully edited !");
+        }else{
+            return redirect()->back();
+        }
     }
 
     /**
@@ -154,9 +173,12 @@ class UsersCRUDcontrol extends Controller
      * @param  \App\User  $user
      * @return \Illuminate\Http\Response
      */
-    public function destroy(User $user)
-    {
-        $user->delete();
-        return redirect('/adashboard/users')->with('success', "Data ->{$user->username}<- was successfully deleted !");
+    public function destroy(User $user){
+        if (Auth::user()->id_level == 1) {
+            $user->delete();
+            return redirect('/adashboard/users')->with('success', "Data ->{$user->username}<- was successfully deleted !");
+        }else{
+            return redirect()->back();
+        }
     }
 }
